+++ tests/test_environment.py
@@ -33,20 +33,18 @@ def test_reset_file_descriptors(pyscript):
     with open(pidfile, 'rb') as f:
         proc = psutil.Process(int(f.read()))
 
-    assert proc.num_fds() > 4
-    open_files = [os.path.relpath(x.path, script.dirname)
-                  for x in proc.open_files()]
-    assert open_files == ['foo.txt', 'foo.pid']
+    open_files = {os.path.relpath(x.path, script.dirname)
+                  for x in proc.open_files()}
+    assert open_files == {'foo.txt', 'foo.pid'}
 
     script.run('restart', '1')
 
     with open(pidfile, 'rb') as f:
         proc = psutil.Process(int(f.read()))
 
-    assert proc.num_fds() == 4
-    open_files = [os.path.relpath(x.path, script.dirname)
-                  for x in proc.open_files()]
-    assert open_files == ['foo.pid']
+    open_files = {os.path.relpath(x.path, script.dirname)
+                  for x in proc.open_files()}
+    assert open_files == {'foo.pid'}
 
     script.run('stop')
 
@@ -82,15 +80,17 @@ def test_chrootdir(pyscript):
 
     # The chroot messes up coverage
     orig_cov_file = os.environ.get('COV_CORE_DATAFILE')
-    cov_file_prefix = os.path.basename(orig_cov_file)
-    os.environ['COV_CORE_DATAFILE'] = '/' + cov_file_prefix
+    if orig_cov_file:
+        cov_file_prefix = os.path.basename(orig_cov_file)
+        os.environ['COV_CORE_DATAFILE'] = '/' + cov_file_prefix
 
     result = script.run()
 
     # Move coverage files to expected location
-    for cov_file in glob(os.path.join(chrootdir, cov_file_prefix + '*')):
-        shutil.move(cov_file, os.path.dirname(orig_cov_file))
-    os.environ['COV_CORE_DATAFILE'] = orig_cov_file
+    if orig_cov_file:
+        for cov_file in glob(os.path.join(chrootdir, cov_file_prefix + '*')):
+            shutil.move(cov_file, os.path.dirname(orig_cov_file))
+        os.environ['COV_CORE_DATAFILE'] = orig_cov_file
 
     assert result.returncode == 0
     assert result.stderr == b'banana\npGh1XcBKCOwqDnNkyp43qK9Ixapnd4Kd\n'