+++ HISTORY.rst
@@ -1,6 +1,14 @@
 Release History
 ---------------
 
+v1.0.2 (not yet released)
+~~~~~~~~~~~~~~~~~~~
+
+* Fixed bug checking if a stream is a socket on Python 3.8.
+* Closing open files is now more efficient on systems with a very high limit
+  on the number of open files.
+* Improved detection of running inside a container.
+
 v1.0.1 (2016-04-17)
 ~~~~~~~~~~~~~~~~~~~